Revert "Move DriveLocalizer classes out of the actual drive classes"
This reverts commit b7f93b3b3f12f2212f97b9de596dd015d6171a66.
