--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MecanumDriveLocalizer.java
@@ -1,105 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.acmerobotics.roadrunner.DualNum;
-import com.acmerobotics.roadrunner.MecanumKinematics;
-import com.acmerobotics.roadrunner.Rotation2d;
-import com.acmerobotics.roadrunner.Time;
-import com.acmerobotics.roadrunner.Twist2dDual;
-import com.acmerobotics.roadrunner.Vector2d;
-import com.acmerobotics.roadrunner.Vector2dDual;
-import com.acmerobotics.roadrunner.ftc.Encoder;
-import com.acmerobotics.roadrunner.ftc.FlightRecorder;
-import com.acmerobotics.roadrunner.ftc.OverflowEncoder;
-import com.acmerobotics.roadrunner.ftc.PositionVelocityPair;
-import com.acmerobotics.roadrunner.ftc.RawEncoder;
-import com.qualcomm.robotcore.hardware.IMU;
-
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
-import org.firstinspires.ftc.teamcode.messages.MecanumLocalizerInputsMessage;
-
-public class MecanumDriveLocalizer implements Localizer {
-    private final MecanumDrive mecanumDrive;
-    public final Encoder leftFront, leftBack, rightBack, rightFront;
-    public final IMU imu;
-
-    private int lastLeftFrontPos, lastLeftBackPos, lastRightBackPos, lastRightFrontPos;
-    private Rotation2d lastHeading;
-    private boolean initialized;
-
-    public MecanumDriveLocalizer(MecanumDrive mecanumDrive) {
-        this.mecanumDrive = mecanumDrive;
-        leftFront = new OverflowEncoder(new RawEncoder(mecanumDrive.leftFront));
-        leftBack = new OverflowEncoder(new RawEncoder(mecanumDrive.leftBack));
-        rightBack = new OverflowEncoder(new RawEncoder(mecanumDrive.rightBack));
-        rightFront = new OverflowEncoder(new RawEncoder(mecanumDrive.rightFront));
-
-        imu = mecanumDrive.lazyImu.get();
-
-        // TODO: reverse encoders if needed
-        //   leftFront.setDirection(DcMotorSimple.Direction.REVERSE);
-    }
-
-    @Override
-    public Twist2dDual<Time> update() {
-        PositionVelocityPair leftFrontPosVel = leftFront.getPositionAndVelocity();
-        PositionVelocityPair leftBackPosVel = leftBack.getPositionAndVelocity();
-        PositionVelocityPair rightBackPosVel = rightBack.getPositionAndVelocity();
-        PositionVelocityPair rightFrontPosVel = rightFront.getPositionAndVelocity();
-
-        YawPitchRollAngles angles = imu.getRobotYawPitchRollAngles();
-
-        FlightRecorder.write("MECANUM_LOCALIZER_INPUTS", new MecanumLocalizerInputsMessage(
-                leftFrontPosVel, leftBackPosVel, rightBackPosVel, rightFrontPosVel, angles));
-
-        Rotation2d heading = Rotation2d.exp(angles.getYaw(AngleUnit.RADIANS));
-
-        if (!initialized) {
-            initialized = true;
-
-            lastLeftFrontPos = leftFrontPosVel.position;
-            lastLeftBackPos = leftBackPosVel.position;
-            lastRightBackPos = rightBackPosVel.position;
-            lastRightFrontPos = rightFrontPosVel.position;
-
-            lastHeading = heading;
-
-            return new Twist2dDual<>(
-                    Vector2dDual.constant(new Vector2d(0.0, 0.0), 2),
-                    DualNum.constant(0.0, 2)
-            );
-        }
-
-        double headingDelta = heading.minus(lastHeading);
-        Twist2dDual<Time> twist = mecanumDrive.kinematics.forward(new MecanumKinematics.WheelIncrements<>(
-                new DualNum<Time>(new double[]{
-                        (leftFrontPosVel.position - lastLeftFrontPos),
-                        leftFrontPosVel.velocity,
-                }).times(MecanumDrive.PARAMS.inPerTick),
-                new DualNum<Time>(new double[]{
-                        (leftBackPosVel.position - lastLeftBackPos),
-                        leftBackPosVel.velocity,
-                }).times(MecanumDrive.PARAMS.inPerTick),
-                new DualNum<Time>(new double[]{
-                        (rightBackPosVel.position - lastRightBackPos),
-                        rightBackPosVel.velocity,
-                }).times(MecanumDrive.PARAMS.inPerTick),
-                new DualNum<Time>(new double[]{
-                        (rightFrontPosVel.position - lastRightFrontPos),
-                        rightFrontPosVel.velocity,
-                }).times(MecanumDrive.PARAMS.inPerTick)
-        ));
-
-        lastLeftFrontPos = leftFrontPosVel.position;
-        lastLeftBackPos = leftBackPosVel.position;
-        lastRightBackPos = rightBackPosVel.position;
-        lastRightFrontPos = rightFrontPosVel.position;
-
-        lastHeading = heading;
-
-        return new Twist2dDual<>(
-                twist.line,
-                DualNum.cons(headingDelta, twist.angle.drop(1))
-        );
-    }
-}
